Notify if build fails

+++ launch.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+if ! make -C "$DIR"; then
+    notify-send "ebar compilation failed"
+    exit 1
+fi
+
 killall ebar
-make -C "$DIR"
 "$DIR/build/ebar" &
